Allow removing / replacing undefined array elements
The following operations currently fail:

  jsonpatch.apply([,1], {"op":"replace","path":"/0","value":0})
  jsonpatch.apply([,1], {"op":"remove","path":"/0"})

When converted to JSON, the array would turn into [null,1] and the above
operations would succeed.

However, it is a common use case to apply patches to JavaScript objects
which are not proper JSON, and it seems reasonable to expect jsonpatch
to treat objects like their JSON equivalents in that case.  In
particular, jsonpatch should allow to remove and replace undefined array
elements.  This commit enables that.